Check the deviceManager
