+++ MM3DM2/ECS/System.swift
@@ -7,5 +7,6 @@
 
 import Foundation
 protocol System {
-    
+    /**I will need to think about Bing's suggestion**/
+    //func update()
 }
